Lock compatibility information if it came from `gecko_android`
Manifest should be the source of truth for compatibility data for Android when
`gecko_android` if set.

+++ src/olympia/devhub/views.py
@@ -1045,11 +1045,13 @@ def version_edit(request, addon_id, addon, version_id):
             timer.log_interval('2.form_validated')
         if 'compat_form' in data:
             for compat in data['compat_form'].save(commit=False):
+                compat.originated_from = amo.APPVERSIONS_ORIGINATED_FROM_DEVELOPER
                 compat.version = version
                 compat.save()
 
             for compat in data['compat_form'].deleted_objects:
-                compat.delete()
+                if not compat.locked_from_manifest:
+                    compat.delete()
 
             for form in data['compat_form'].forms:
                 if isinstance(form, forms.CompatForm) and 'max' in form.changed_data:
